k-gram Probability Smoothers

Information on available k-gram continuation probability smoothers.

List of smoothers currently supported by kgrams

"ml": Maximum Likelihood estimate.

"add_k": Add-k smoothing.

"laplace": Laplace (A.K.A. Add-one) smoothing.

"kn": Interpolated Kneser-Ney with fixed discount. (Kneser and Ney 1995; Chen and Goodman 1999) .

"sbo": Stupid Backoff (Brants et al. 2007) .

smoothers()

info(smoother)

Arguments

smoother
a string. Code name of probability smoother.

Value

smoothers() returns a character vector, the list of code names of probability smoothers available in kgrams. info(smoother) returns NULL and prints some information on the selected smoothing technique.

References

Brants T, Popat AC, Xu P, Och FJ, Dean J (2007). “Large Language Models in Machine Translation.” In Proceedings of the 2007 Joint Conference on Empirical Methods in Natural Language Processing and Computational Natural Language Learning (EMNLP-CoNLL), 858--867. https://www.aclweb.org/anthology/D07-1090.

Chen SF, Goodman J (1999). “An empirical study of smoothing techniques for language modeling.” Computer Speech \& Language, 13(4), 359--394.

Kneser R, Ney H (1995). “Improved backing-off for M-gram language modeling.” 1995 International Conference on Acoustics, Speech, and Signal Processing, 1, 181--184.

Author

Valerio Gherardi

Examples

# List available smoothers
smoothers()

#> [1] "sbo"     "add_k"   "laplace" "ml"      "kn"     

# Get information on smoother "kn", i.e. Interpolated Kneser-Ney
info("kn")

#> Interpolated Kneser-Ney with fixed discount
#>  * code: 'kn'
#>  * parameters: D (discount in higher order probability part)
#>  * constraints: 0 < D < 1