k-gram Probability Smoothers

Information on available k-gram continuation probability smoothers.

List of smoothers currently supported by kgrams

"ml": Maximum Likelihood estimate markov1913essaikgrams.

"add_k": Add-k smoothing dale1995philosophical,lidstone1920note,johnson1932probability,jeffreys1998theorykgrams.

"abs": Absolute discounting ney1991smoothingkgrams.

"wb": Witten-Bell smoothing bell1990text,witten1991zerokgrams

"kn": Interpolated Kneser-Ney. Kneser1995ImprovedBF,chen1999empiricalkgrams.

"mkn": Interpolated modified Kneser-Ney. chen1999empiricalkgrams.

"sbo": Stupid Backoff brants-etal-2007-largekgrams.

smoothers()

info(smoother)

Arguments

smoother

a string. Code name of probability smoother.

Value

smoothers() returns a character vector, the list of code names of probability smoothers available in kgrams. info(smoother) returns NULL (invisibly) and prints some information on the selected smoothing technique.

References

Author

Valerio Gherardi

Examples

# List available smoothers
smoothers()
#> [1] "ml"    "add_k" "abs"   "kn"    "mkn"   "sbo"   "wb"   

# Get information on smoother "kn", i.e. Interpolated Kneser-Ney
info("kn")
#> Interpolated Kneser-Ney
#>  * code: 'kn'
#>  * parameters: D
#>  * constraints: 0 <= D <= 1